# backend/swagger.yaml
openapi: 3.0.0
info:
  title: Smart Task Manager API
  version: 1.0.0
paths:
...
  /api/tasks:
    get:
      summary: Get tasks
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Tasks list }
    post:
      summary: Create task
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Task' }
      responses:
        '201': { description: Task created }
...
components:
  schemas:
    Task:
      type: object
      properties:
        title: { type: string }
        description: { type: string }
        status: { type: string, enum: [todo, in_progress, completed] }
        priority: { type: string, enum: [low, medium, high] }
        deadline: { type: string, format: date-time }
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT